+++ src/components/posts/PostDetails.js
@@ -8,7 +8,7 @@ import { HumanDate } from "../utils/HumanDate";
 export const PostDetails = ({ token}) => {
   const [post, setPost] = useState({})
   const [user, setUser] = useState({})
-    const { userId } = useParams()
+  const { userId } = useParams()
 
   const { postId } = useParams()
